requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
PyMySQL==1.1.2
python-dotenv==1.2.1
Werkzeug==3.1.4
